Added auto-aim, gun assets, and updated the weapon model
The player can now activate auto-aim by pressing 'R' which finds the nearest enemy to the player and then rotates to face that enemy while continuously shooting. If auto-aim is disabled the player still has to manually turn and shoot.

Added a low poly gun pack and changed the model for the players weapon.