Game over screen and bug fixes
Added a game over screen when the player dies and they are given an option to restart or quit. Currently quitting doesn't do anything which is the same as the pause menu.
Fixed some bugs with animations happening whilst menus are open.
Fixed the bug where projectiles would be initialized whilst the game is paused.