Refactor enemy spawner
Refactored the enemy spawner and enemy class to account for different types of enemies easier.
This allows adding more types of enemies without different tags as this makes other systems like the amount of enemies killed not function correctly.
This also allows for different attacks to   be added easier as well.