Fixed
Fixed gem bounce/position.
Added NavMesh agents to the player and all enemies.
Added a NavMesh surface to the level.
Made the crate a NavMesh obstacle.
Changed the rake to melee and added a boss rake. Fixed the animations of the rake.
The animation controller for the rake/catfish are now separate and named appropriately. Added an animation controllers folder.