Speed buff icon
Added a speed buff icon to the bottom middle of the screen whenever the player picks up a buff.
Changed how it works so the player cannot pick up multiple speed buffs. If the player picks up a speed whilst one is already active then the duration of it is reset.